Refactored code
Basically, I just swapped the axes of the dictionary. In this version, the number of rows correspond to the number of atoms, and the number of columns correspond to the length of the original signal. This allows me to  use one less `np.ascontiguousarray` call in the OMP algorithm. I also had to make the necessary adjustments in all other functions. I also changed `np.full` to `np.empty` because it's not like I need the values to be set to some initial value. Overall, the idea behind this commit is to optimize the code.